refactor: Emoticon 모델 생성
