+++ EmoticonApp/EmoticonApp/Data/Entity/Emoticon.swift
@@ -1,0 +1,14 @@
+//
+//  Emoticon.swift
+//  EmoticonApp
+//
+//  Created by 지현우 on 2021/01/20.
+//
+
+import Foundation
+
+struct Emoticon{
+    let title: String
+    let author: String
+    let imagePath: String
+}
